+++ extensions/iceworks-project-creator/web/src/pages/CreateProject/components/ScaffoldMarket/index.tsx
@@ -118,6 +118,8 @@ const ScaffoldMarket = ({ onScaffoldSelect, curProjectField, children, onOpenCon
           {loading ? <Loading visible={loading} className={styles.loading} /> : <>
             <div className={styles.mainScaffolds}>
               {!!mainScaffolds.length ? mainScaffolds.map(item => {
+                // tsScaffoldsList and jsScaffoldsList only contain the official scaffolds
+                // so the TypeScript and JavaScript logo only display in official scaffolds
                 const scaffoldType = tsScaffoldsList.includes(item.source.npm) ? 'ts' :
                   jsScaffoldsList.includes(item.source.npm) ? 'js' :
                     '';
@@ -129,7 +131,7 @@ const ScaffoldMarket = ({ onScaffoldSelect, curProjectField, children, onOpenCon
                     title={
                       <div className={styles.cardTitle}>
                         {scaffoldType && <img src={require(`@/assets/${scaffoldType}.svg`)} alt="languageType" width={20} height={20} />}
-                        <div>{item.title.replace(' - TS', '')}</div>
+                        <div>{scaffoldType ? item.title.replace(' - TS', '').replace(' - JS', '') : item.title}</div>
                       </div>
                     }
                     content={item.description}
@@ -146,6 +148,8 @@ const ScaffoldMarket = ({ onScaffoldSelect, curProjectField, children, onOpenCon
               <Collapse.Panel title="查看更多">
                 <div className={styles.collapseScaffolds}>
                   {otherScaffolds.map(item => {
+                    // tsScaffoldsList and jsScaffoldsList only contain the official scaffolds
+                    // so the TypeScript and JavaScript logo only display in official scaffolds
                     const scaffoldType = tsScaffoldsList.includes(item.source.npm) ? 'ts' :
                       jsScaffoldsList.includes(item.source.npm) ? 'js' :
                         '';
@@ -157,7 +161,7 @@ const ScaffoldMarket = ({ onScaffoldSelect, curProjectField, children, onOpenCon
                         title={
                           <div className={styles.cardTitle}>
                             {scaffoldType && <img src={require(`@/assets/${scaffoldType}.svg`)} alt="languageType" width={20} height={20} />}
-                            <div>{item.title.replace(' - JS', '')}</div>
+                            <div>{scaffoldType ? item.title.replace(' - JS', '').replace(' - TS', '') : item.title}</div>
                           </div>
                         }
                         content={item.description}